requirements.txt
streamlit
pandas
python-dotenv
openai
pydantic
langchain
langchain-openai
langchain-community
faiss-cpu
plotly
